+++ core/hw/sh4/sh4_sched.h
@@ -50,6 +50,7 @@ void sh4_sched_ffts();
 */
 void sh4_sched_idle_fastforward();
 void DYNACALL sh4_delay_loop_skip(u32 pc, u32 cyc);
+void DYNACALL sh4_sched_idle_fastforward_if_ram(u32 reg);
 extern u64 g_idleFFCalls, g_idleFFCycles;
 
 struct sched_list